+++ folder/entrie_class.py
@@ -1,0 +1,32 @@
+from tkinter import Entry
+# class for entries storage (класс для хранения текстовых полей)
+class Entries:
+    def __init__(self):
+        self.entries_list = []
+        self.parent_window = None
+
+    def set_parent_window(self, parent_window):
+        self.parent_window = parent_window
+
+    # adding of new entry (добавление нового текстового поля)
+    def add_entry(self):
+        new_entry = Entry(self.parent_window)
+        new_entry.icursor(0)
+        new_entry.focus()
+        new_entry.pack()
+        new_entry.focus_get()
+        plot_button = self.parent_window.get_button_by_name('plot')
+        if plot_button:
+            plot_button.pack_forget()
+        self.parent_window.add_button('plot', 'Plot', 'plot', hot_key='<Return>')
+        self.entries_list.append(new_entry)
+
+    def add_filled_entry(self, text):
+        new_entry = Entry(self.parent_window)
+        new_entry.insert(0, text)
+        new_entry.pack()
+        plot_button = self.parent_window.get_button_by_name('plot')
+        if plot_button:
+            plot_button.pack_forget()
+        self.parent_window.add_button('plot', 'Plot', 'plot', hot_key='<Return>')
+        self.entries_list.append(new_entry)
